+++ honeycomb/__init__.py
@@ -2,9 +2,10 @@ from .hive import run_lake_query
 from .append_table import append_df_to_table
 from .create_table import create_table_from_df, flash_update_table_from_df
 from .describe_table import describe_table
-
 from .meta import get_table_storage_type_from_metadata
 from . import alter_table, check
+from . import bigquery, salesforce
+
 
 __all__ = [
     'alter_table',
@@ -15,4 +16,6 @@ __all__ = [
     'create_table_from_df',
     'describe_table',
     'get_table_storage_type_from_metadata',
+    'bigquery',
+    'salesforce'
 ]